hpsc: hpps: dt: exclude Linux ramoops buf
The memory for Linux rammops is excluded from the total avalable memory
(like memory for ATF is excluded).

It needs to be excluded because otherwise U-boot zeros its contents, but
we want to emulate a persistent buffer (that is, pre-load the ramoops
with contents using HW emulator, e.g. Qemu, so that its contents can be
decoded).

Excluded memory is not visible to Linux, unless Linux declares it as a
reserve memory region.  In this case, we do explicitly reserve the
memory for the ramoops subsystem in Linux device tree.

Conflicts:
	arch/arm/dts/hpsc-hpps.dts