hpsc: hpps: dt: expose mem for a GP buffer
Linux reserve this region to use a general-purpose mechanism
to get data out of the system without requiring any I/O device
(but only memory dump).

As far as I can tell, declaring that memory region available here does
not lead U-boot to use this region for anything (we do not want that).
It is for sure not used by Linux (for any of its internal purposes),
because we declare the region as reserved in Linux DT.

Conflicts:
	arch/arm/dts/hpsc-hpps.dts