hpsc: hpps: dt: move all regions to first 64 Mb
HW emulators don't have the maximum amount of DDR memory,
so keep everything as tight as possible.

The general-purpose buffer is outside of first 64Mb, but it is
non-essential -- it is not needed to boot the system.  Also, the gp
buffer can easily be dropped, without affecting any other regions.