hpsc: cfg: disable ramdisk relocation
With the current memory layout (see DT), where HPPS DDR LOW memory is
split into two regions, relocation fails to allocate memory for a large
ramdisk (e.g.  >32MB, in the use case where the whole rootfs is loaded
as an initramfs). It is not clear why it fails, but it is most likely
due to allocator not being able to handle more than one fragment, and
the first fragment alone is certainly too small for a large ramdisk.