hpsc: hpps: cfg: u-boot DT address in early env
By default u-boot DT is placed after the u-boot image, which may fall
onto an address that is not aligned to DMA burst size, which precludes
copying it via DMA (or introduces complexity due to non-uniform busrt
size across images, or non-uniform copy method accross images).

So, we just take control over this address by defining it in early env.

Note that this address is not overridable in (late) environment.